Коммит №12.1: refactor: -Переработан метод delete в классах FilmDbStorage и FilmService; -Переработан метод delete в классе UserDbStorage и UserService; -Испаравлена ER-диаграмма.
-fix:
-Исправлена ошибка в методе DeleteById в классе FilmController;
-Исправлена ошибка в методе DeleteById в классе UserController.

Style:
-Косметические изменения по всему проекту.

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -52,12 +52,19 @@ public class FilmService {
         return result;
     }
 
-    public void delete(int filmId) {
+    public Film delete(Integer filmId) {
+        log.info("Request to delete the movie by ID = " + filmId + " received");
+        if (filmId == null) {
+            throw new NotFoundException("Movie with ID = " + filmId + " not found");
+        }
+        if (filmId < 0) {
+            throw new NotFoundException("Movie with ID = " + filmId + " not found");
+        }
         if (getById(filmId) == null) {
             throw new NotFoundException("Movie with ID = " + filmId + " not found");
         }
         log.info("Deleted film with id: {}", filmId);
-        filmStorage.delete(filmId);
+        return filmStorage.delete(filmId);
     }
 
     public Film getById(Integer id) {
